backend/requirements.txt
fastapi
uvicorn
pillow
numpy
pydantic
mediapipe
